+++ python/paddle/v2/framework/tests/test_lookup_table_op.py
@@ -11,7 +11,7 @@ class TestLookupTableOp(OpTest):
     def setUp(self):
         self.op_type = "lookup_table"
         table = np.random.random((17, 31)).astype("float32")
-        ids = np.random.randint(0, 17, 4).astype("int32")
+        ids = np.random.randint(0, 17, 4).astype("int64")
         self.inputs = {'W': table, 'Ids': ids}
         self.outputs = {'Out': table[ids]}
